refactor(core/rust/crypto: introduce testing harness and zkp support
moves the test-only `rand` insecure PRNG into a dedicated test-only
module `testutil`

implements `random_buffer` extern "C" export that delegates to the
insecure PRNG
and applies a xor with a masking string <PRNG-Rust-Tests> so that its
presence can be detected within a compiled artifact
(required for secp256-zkp EC arithmetic even for verification because
curve_to_jacobian needs random data for FIH masking)

explicitly runs the test harness and initializes the zkp context before
start (also required by secp256-zkp)

moves the secp256-zkp default error handlers from modtrezorcrypto.c to
the crypto crate, so that they're properly set for any user of crypto,
not just firmware.
(this removes the "raise ValueError" feature for illegal arguments --
that shouldn't matter too much, these are checks on "is the function
called properly", not on validity of user provided input)
