+++ core/embed/upymod/modtrezorcrypto/modtrezorcrypto.c
@@ -167,17 +167,4 @@ const mp_obj_module_t mp_module_trezorcrypto = {
 
 MP_REGISTER_MODULE(MP_QSTR_trezorcrypto, mp_module_trezorcrypto);
 
-#ifdef USE_SECP256K1_ZKP
-void secp256k1_default_illegal_callback_fn(const char *str, void *data) {
-  (void)data;
-  mp_raise_ValueError((mp_rom_error_text_t)str);
-  return;
-}
-
-void secp256k1_default_error_callback_fn(const char *str, void *data) {
-  (void)data;
-  error_shutdown(str);
-}
-#endif
-
 #endif  // MICROPY_PY_TREZORCRYPTO
